double latitude longitude

+++ src/main/java/com/bandtec/techconnective/model/Geolocalizacao.java
@@ -1,31 +1,29 @@
 package com.bandtec.techconnective.model;
 
-import java.util.Arrays;
-
 public class Geolocalizacao {
-    private Integer latitude;
-    private Integer longitude;
+    private Double latitude;
+    private Double longitude;
 
-    public Geolocalizacao(Integer latitude, Integer longitude) {
+    public Geolocalizacao(Double latitude, Double longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
     public Geolocalizacao(){}
 
-    public Integer getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Integer latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Integer getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Integer longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
